exception divide 0, results exceeds 100

+++ src/main/java/Calculating.java
@@ -5,6 +5,6 @@ public interface Calculating {
     double add(double a, double b);
     double subtract(double a, double b);
     double multiply(double a, double b);
-    double divide(double a, double b);
+    int divide(int a, int b);
 }
